Rename tests/parallel_runner.py -> tests/parallel_testsuite.py
This file defines a parallel test suite.  We always use
unittest.TextTestRunner to actually run the suites.
